6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,8 @@
 #include <stdio.h>
 /**
-* * Description: main - A program that print the size of various types on the computer
-* * Return 0 (Success)
+* main - Entry point
+*
+* Return 0 (Success)
 */
 int main(void)
 {
@@ -10,10 +11,10 @@ int b;
 long int c;
 long long int d;
 float f;
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
-return (0);
+	printf("Size of a char: %d byte(s)\n", sizeof(a));
+	printf("Size of a int: %d byte(s)\n", sizeof(b));
+	printf("Size of a long int: %d byte(s)\n", sizeof(c));
+	printf("Size of a long long int: %d byte(s)\n", sizeof(d));
+	printf("Size of a float: %d byte(s)\n", sizeof(f));
+	return (0);
 }
